requirements.txt
numpy==1.19.2
opencv_python==3.4.3.18
Pillow==8.4.0
scikit_learn==1.0.1
streamlit==1.1.0
tensorflow==2.4.0
